Finished preprocessing
moved logregsgd file into the src
finished and changed some of the preprocessing commands to start working on building the models made sure to keep cloud columns by replacing na it -1
changed for loop to and used a pandas command to get the mean finished getting all the averages for the rest of the columns. Made sure that the data types are the same. added some spaces to make code easier to read for me